requirements.txt
pyaudio==0.2.13
openai-whisper==20231117
torch==2.1.1
webrtcvad==2.0.10
numpy==1.24.3
